+++ tcod/cffi.h
@@ -13,4 +13,5 @@
 #include "noise.h"
 #include "path.h"
 #include "random.h"
+#include "tcod.h"
 #include "tdl.h"