+++ tcod/tcod.h
@@ -1,0 +1,7 @@
+
+#ifndef TCOD_TCOD_H_
+#define TCOD_TCOD_H_
+
+int LineWhere(int x1, int y1, int x2, int y2, int *x_out, int *y_out);
+
+#endif /* TCOD_TCOD_H_ */